document-service/requirements.txt
fastapi>=0.115
python-multipart>=0.0.9
uvicorn[standard]>=0.30
pydantic>=2.8
openpyxl>=3.1
pypdf>=4.3
python-docx>=1.1
beautifulsoup4>=4.12
markdown>=3.6
charset-normalizer>=3.3
pytest>=8.3
httpx>=0.27

